+++ candle-metal-kernels/src/kernels/mod.rs
@@ -1,11 +1,13 @@
 pub mod binary;
 mod macros;
 pub mod mlx_gemm;
+pub mod sdpa;
 pub mod sort;
 pub mod unary;
 
 pub use binary::{call_binary_contiguous, call_binary_strided};
 pub use mlx_gemm::{call_mlx_gemm, GemmDType};
+pub use sdpa::{call_sdpa_full, call_sdpa_vector, call_sdpa_vector_2pass};
 pub use sort::{call_arg_sort, call_mlx_arg_sort};
 pub use unary::{
     call_const_set_contiguous, call_const_set_contiguous_tiled, call_const_set_strided,
